Remove `sqla` module from security manager
